Start work on top media editor
